0.0.117 - data.model and view.model leads to breakthrough with controls example 11b - shared data model being used to mirror the value of 2 text fields
Still would benefit from low/mid level code to persist or reconstruct the data models on the client, or isomorphic construction for them on both the client and server

The dual view.model and data.model system is the beginning of a more flexible multi-model system, where controls can share a model, while also themselves being in a different state.

Also consider using this model system to enable user, session, or control specific settings, also enabling it so they could be transferred to other users as screen sharing, but that such settings / options of the user (and other) specific levels are distinct from other levels of settings / options, such as the way the app is designed.